Extract round-card/gate-media MCP tools out of mcp-server.ts (Wave C batch 5)
mcp-server.ts 3976 -> 3977 (adjusted for spread lines) from 4578 lines.
Splits show_round/show_media/get_round_status into
mcp-round-card-tools.ts (createRoundCardTools) and
get_gate_verdict/get_round_media/get_gate_media into
mcp-gate-media-tools.ts (createGateMediaTools).

show_round/show_media used to call tools.get_round_status.handler(...)
as a self-reference into the tools object; that becomes a local
roundStatusHandler function shared by both callers and wired directly
as get_round_status's handler. resolveAuth is injected with a wider
return shape for the round-card half (record/access/issueNumber, not
just channelToken); store and now are newly injected deps.
