+++ apps/api/src/agent-surface/mcp-server.ts
@@ -1,7 +1,7 @@
 import type { FastifyInstance, FastifyReply, FastifyRequest } from 'fastify';
 import { createHash, timingSafeEqual } from 'node:crypto';
 import { z } from 'zod';
-import { AGENT_CHANNEL_ROUTES, GATE_STATUS_VALUES, type BuilderKind } from '@gamedevpl/contract';
+import { AGENT_CHANNEL_ROUTES, type BuilderKind } from '@gamedevpl/contract';
 import {
   toolOk,
   toolErr,
@@ -23,6 +23,8 @@ import { createKitTools } from './mcp-kit-tools.js';
 import { createKitFileTools } from './mcp-kit-file-tools.js';
 import { createInboxTools } from './mcp-inbox-tools.js';
 import { createSeedTools } from './mcp-seed-tools.js';
+import { createRoundCardTools } from './mcp-round-card-tools.js';
+import { createGateMediaTools } from './mcp-gate-media-tools.js';
 
 export { MCP_VISIBLE_TOOLS };
 import { looksLikeCreatorAgentKey } from './agent-creator-key.js';
@@ -79,7 +81,6 @@ import {
 } from '../delivery/games-store.js';
 import { deriveGateStatusString, readGateVerdict } from '../delivery/gate-verdict.js';
 import { gameManifestHint } from './game-manifest-hint.js';
-import { detectStall, resolveJobState, toSubmissionStatus } from '../creation/job-state.js';
 import { largeSourceFileHint, moduleSizeWarnings } from '../creation/module-size.js';
 import type { GcsObjectStore } from '../delivery/gcs-sign.js';
 import {
@@ -176,101 +177,6 @@ const MAX_MCP_BODY_BYTES = 2 * 1024 * 1024;
 const MAX_SUBMIT_FILES = MAX_UPLOAD_FILES;
 const MAX_STAGE_UPLOAD_BATCH = 50;
 
-/** How often the round view should re-read status. Matches the gate's own backoff. */
-/**
- * The progress note in the language the *reader* is using.
- *
- * Same rule as the creator's thread (`submissions.ts`): an agent may send both an English
- * `text` and a `textLocalized` in the creator's language, and the localized one is shown
- * only when its locale matches the reader's. The card used to prefer `textLocalized`
- * unconditionally, which handed a Polish note to a reader whose surface was English —
- * the same event rendered in two languages depending on where you looked (owner, 2026-08-05).
- *
- * Locales are compared on the primary subtag, so `pl` matches `pl-PL`. No locale from the
- * host means no claim about the reader, so English is the safe answer.
- */
-function noteTextFor(event: { text: string; textLocalized?: string; locale?: string }, readerLocale: unknown): string {
-  if (!event.textLocalized || typeof event.locale !== 'string') return event.text;
-  if (typeof readerLocale !== 'string' || !readerLocale) return event.text;
-  const primary = (value: string) => value.trim().toLowerCase().split(/[-_]/)[0];
-  return primary(event.locale) === primary(readerLocale) ? event.textLocalized : event.text;
-}
-
-/**
- * The round snapshot both doors return: `show_round` for the model, `get_round_status`
- * for the card. Shared so the two can never drift into describing different rounds.
- */
-const ROUND_STATUS_OUTPUT_SCHEMA: Record<string, unknown> = {
-  type: 'object',
-  properties: {
-    phase: { type: 'string', description: 'Internal job state.' },
-    status: { type: 'string', description: 'Creator-facing projection of the phase.' },
-    stall: { type: ['string', 'null'] },
-    agentEnded: { type: 'boolean' },
-    title: { type: ['string', 'null'] },
-    slug: { type: ['string', 'null'] },
-    round: { type: 'number' },
-    deliveriesRemaining: { type: ['number', 'null'] },
-    note: {
-      type: ['object', 'null'],
-      properties: { text: { type: 'string' }, createdAt: { type: 'string' } },
-    },
-    shot: {
-      type: ['object', 'null'],
-      properties: {
-        id: { type: 'string' },
-        createdAt: { type: 'string' },
-        label: { type: ['string', 'null'] },
-        png: { type: 'string', description: 'base64 PNG; omitted when unchanged since sinceShotId.' },
-      },
-    },
-    gate: { type: ['object', 'null'] },
-    retryAfterSeconds: { type: 'number' },
-  },
-  required: ['phase', 'status', 'retryAfterSeconds'],
-};
-
-/**
- * Where the creator plays what the agent just built.
- *
- * Built here rather than in the view because the view is served to every environment
- * from one string: an origin baked into it would send a staging card's Play button to
- * production.
- *
- * `canonicalAppBaseUrl()` rather than `WEB_ORIGIN`, and the distinction is load-bearing.
- * WEB_ORIGIN is a CORS allowlist — in production it begins with the Cloud Run service
- * URL — so taking its first entry produced a link to an origin that is not in the view's
- * `redirect_domains`, which ChatGPT would refuse. The button would have been dead in
- * production and fine everywhere we tested it (Codex, #617). The card's link allowlist
- * is derived from this same function so the two cannot drift apart again.
- */
-function playUrlFor(slug: string | null | undefined): string | null {
-  if (!slug) return null;
-  return `${canonicalAppBaseUrl()}/play/${encodeURIComponent(slug)}`;
-}
-
-/**
- * The round's home in Creator Studio — where the creator manages this build.
- *
- * The card's title links here rather than to /play, which the Play button already
- * covers. It is also always valid: Studio shows a round whether or not anything is
- * playable yet, so it has none of the "not available" risk that keeps the Play button
- * off a red gate.
- */
-function studioUrlFor(slug: string | null | undefined): string | null {
-  if (!slug) return null;
-  return `${canonicalAppBaseUrl()}/studio/${encodeURIComponent(slug)}`;
-}
-
-/** The site itself, for the card's wordmark. */
-function siteUrl(): string {
-  return canonicalAppBaseUrl();
-}
-
-const ROUND_STATUS_RETRY_AFTER_SECONDS = 30;
-/** A card shows a strip, not a contact sheet; the bytes ride a postMessage. */
-const ROUND_MEDIA_MAX_FRAMES = 3;
-const ROUND_MEDIA_BYTE_BUDGET = 1_500_000;
 const TRANSPORT_SESSION_TTL_MS = 24 * 60 * 60 * 1000;
 const MAX_TRANSPORT_SESSIONS = 10_000;
 
@@ -3655,516 +3561,8 @@ export async function registerMcpServerRoutes(app: FastifyInstance, options: Mcp
       },
     },
 
-    /**
-     * The only tool that opens the round card, and its whole purpose.
-     *
-     * The card used to hang off `start` and `get_gate_verdict`. That made rendering a
-     * *side effect* of doing something else, so agent behaviour we neither control nor
-     * should have to decided how many cards appeared: an agent that re-ran `start`
-     * before each operation left one card per call (ChatGPT, 2026-08-05). No amount of
-     * host-side reasoning fixes that, because the agent was not doing anything wrong
-     * enough to forbid.
-     *
-     * Showing the creator something is a deliberate act, so it gets a deliberate tool.
-     * Calling it twice is the agent asking for two cards, which is at least honest.
-     *
-     * It also removes the opening scramble: the card seeds from this result, so it has
-     * the round key from its first frame instead of making a keyless poll and waiting
-     * for `start`'s result to arrive.
-     */
-    show_round: {
-      annotations: { title: 'Show the creator a live round card', ...READS },
-      description:
-        "Render a live status card for this round in the creator's chat: phase, latest progress note and " +
-        'screenshot, gate verdict, deliveries left. It refreshes itself and stops when the round settles, so ' +
-        'the creator can watch without you polling. ' +
-        'Call it ONCE per round, after start — a second call renders a second card. ' +
-        'A preview_failed / red card is not finished: honour warnings.code=must_fix_gate, fix, and ' +
-        'submit_sources again — show_round alone does not re-run the gate. ' +
-        'Only clients that render MCP Apps views see anything; elsewhere it is a plain status read. ' +
-        BEHAVIOURAL_CONTRACT,
-      inputSchema: {
-        type: 'object',
-        properties: { sessionKey: SESSION_KEY_PROP },
-        required: [],
-      },
-      // Same body as get_round_status plus the echoed key — they are the same read, so
-      // they declare the same shape.
-      outputSchema: ROUND_STATUS_OUTPUT_SCHEMA,
-      handler: async (args, ctx) => {
-        // One implementation, two doors: the model opens the card through this tool, the
-        // card refreshes itself through the app-only one. They must never disagree about
-        // what the round looks like, so they are literally the same read.
-        const status = await tools.get_round_status.handler(args, ctx);
-        // Echo the key so the view holds it from its first frame. The card reads it off
-        // this result; without it the first poll goes out keyless and is refused.
-        if (status.isError || typeof args.sessionKey !== 'string') return status;
-        const data = { ...(status.structuredContent as Record<string, unknown>), sessionKey: args.sessionKey };
-        return toolOk(data);
-      },
-    },
-
-    /**
-     * Show the creator the gate's pictures — the only way to actually show them.
-     *
-     * `get_gate_media` attaches frames as image blocks, which reach the *model*: it can
-     * look at them and describe them. There is no path back out. Asked to display them,
-     * ChatGPT answered "the image attachments apparently didn't render in your view"
-     * and the creator saw nothing (owner, 2026-08-06). A tool result is input to a
-     * model, not output to a person.
-     *
-     * A view is the only surface that puts pixels in the conversation without going
-     * through the model, so this is not a nicer way to show a picture — it is the way.
-     *
-     * Deliberately carries no bytes. The card fetches them over the app-only
-     * `get_round_media`, so a megabyte of base64 never enters the model's context — the
-     * same reason that tool exists at all.
-     */
-    show_media: {
-      annotations: { title: "Show the creator the gate's screenshots", ...READS },
-      description:
-        "Render the gate's screenshots for a delivery in the creator's chat, at a size they can actually look at, " +
-        'with the gameplay recording and a link to play. ' +
-        'Use this when the creator asks to see the game — get_gate_media lets *you* look at the frames, but its ' +
-        'attachments are input to you rather than something the creator sees. ' +
-        'Defaults to the latest delivery. Only clients that render MCP Apps views show anything. ' +
-        BEHAVIOURAL_CONTRACT,
-      inputSchema: {
-        type: 'object',
-        properties: {
-          sessionKey: SESSION_KEY_PROP,
-          deliveryId: { type: 'string', description: "Delivery to show; default is the round's latest." },
-        },
-        required: [],
-      },
-      outputSchema: ROUND_STATUS_OUTPUT_SCHEMA,
-      handler: async (args, ctx) => {
-        const status = await tools.get_round_status.handler(args, ctx);
-        if (status.isError) return status;
-        const auth = await resolveAuth(ctx, args, { allowTerminalReceipt: true });
-        const record = 'record' in auth ? auth.record : null;
-        const wanted =
-          (typeof args.deliveryId === 'string' && args.deliveryId.trim()) ||
-          record?.previewVersion ||
-          record?.deliveredVersion ||
-          null;
-        return toolOk({
-          ...(status.structuredContent as Record<string, unknown>),
-          // Names the delivery the card should fetch frames for. It need not belong to
-          // this round: "show me the screenshot" opens a round that has delivered
-          // nothing, and the frames the creator means are the previous one's.
-          mediaDeliveryId: wanted,
-          ...(typeof args.sessionKey === 'string' ? { sessionKey: args.sessionKey } : {}),
-        });
-      },
-    },
-
-    /**
-     * App-only (SEP-1865 `visibility: ["app"]`): the round view calls this, the model
-     * never sees it. Read-only and presence-neutral by construction — see
-     * MCP_UI_APP_ONLY_TOOLS for why both matter.
-     */
-    get_round_status: {
-      annotations: { title: 'Round status for the round view', ...READS },
-      outputSchema: ROUND_STATUS_OUTPUT_SCHEMA,
-      description:
-        'Round state for the gamedev.pl round view: phase, latest progress note, latest screenshot and the current ' +
-        'gate verdict in one read. Intended for the view, which polls it on retryAfterSeconds — pass sinceShotId to ' +
-        'skip re-sending screenshot bytes you already hold. Read-only, and deliberately does not count as agent ' +
-        'presence: a creator watching does not keep a quiet round looking alive.',
-      inputSchema: {
-        type: 'object',
-        properties: {
-          sessionKey: SESSION_KEY_PROP,
-          locale: {
-            type: 'string',
-            description:
-              "The reader's language (BCP-47), so a localized progress note is shown only to who can read it.",
-          },
-          sinceShotId: {
-            type: 'string',
-            description: 'Screenshot id you already have; bytes are omitted when the latest shot still matches it.',
-          },
-        },
-        required: [],
-      },
-      handler: async (args, ctx) => {
-        const auth = await resolveAuth(ctx, args, { allowTerminalReceipt: true });
-        if (!('channelToken' in auth)) return auth;
-        if (!store) return toolErr('the MCP build endpoint is not configured');
-
-        const record = auth.record;
-        const state = resolveJobState(record) ?? 'queued';
-        const stall = detectStall({
-          state,
-          stateSince: record.stateSince ?? new Date(now()).toISOString(),
-          ...(record.lastAgentSignalAt ? { lastAgentSignalAt: record.lastAgentSignalAt } : {}),
-          ...(record.agentEndedAt ? { agentEndedAt: record.agentEndedAt } : {}),
-          ...(record.builder ? { builder: record.builder } : {}),
-          now: now(),
-        });
-
-        const [events, shots] = await Promise.all([
-          store.listBuildEvents(auth.issueNumber, { limit: 1 }),
-          store.listBuildShots(auth.issueNumber, { limit: 1 }),
-        ]);
-
-        const latestEvent = events[0];
-        const latestShot = shots[0];
-        const sinceShotId = typeof args.sinceShotId === 'string' ? args.sinceShotId.trim() : '';
-        let shot: Record<string, unknown> | null = null;
-        if (latestShot) {
-          // Bytes only when the view does not already hold this frame — a shot can be
-          // hundreds of KB of base64 and this is polled for the length of a round.
-          const full = latestShot.id !== sinceShotId ? await store.getBuildShot(auth.issueNumber, latestShot.id) : null;
-          shot = {
-            id: latestShot.id,
-            createdAt: latestShot.createdAt,
-            label: latestShot.labelLocalized ?? latestShot.label ?? null,
-            ...(full ? { png: full.data } : {}),
-          };
-        }
-
-        const cap = record.builder === 'self' ? selfBuildDeliveryCap() : null;
-        const used = record.roundDeliveryCount ?? 0;
-
-        // The gate goes through the channel exactly as get_gate_verdict does, so both
-        // read one implementation of what a verdict means — but the channel answers for
-        // the job's latest delivery whatever round produced it. A round that has
-        // delivered nothing has no verdict, and showing the previous round's told the
-        // creator their fresh round had already been refused.
-        //
-        // Terminal receipts are the exception, and the delivery count cannot see it:
-        // closing a round resets that count to 0 (store.ts, `closes`), so a card reading
-        // the green verdict that just closed the round would find nothing and go back to
-        // polling. A receipt is a closed round being read, not a fresh one.
-        let gate: unknown = null;
-        if (used > 0 || auth.access === 'terminal_receipt') {
-          const gateRes = await injectChannel(ctx.request, 'GET', AGENT_CHANNEL_ROUTES.GATE, auth.channelToken);
-          if (gateRes.statusCode === 200) gate = gateRes.json();
-        }
-
-        return toolOk({
-          phase: state,
-          status: toSubmissionStatus(state),
-          stall: stall ?? null,
-          agentEnded: Boolean(record.agentEndedAt),
-          title: record.title ?? null,
-          slug: record.slug ?? null,
-          playUrl: playUrlFor(record.slug),
-          studioUrl: studioUrlFor(record.slug),
-          siteUrl: siteUrl(),
-          round: record.roundGeneration ?? 1,
-          deliveriesRemaining: cap === null ? null : Math.max(0, cap - used),
-          note: latestEvent ? { text: noteTextFor(latestEvent, args.locale), createdAt: latestEvent.createdAt } : null,
-          shot,
-          gate,
-          retryAfterSeconds: ROUND_STATUS_RETRY_AFTER_SECONDS,
-        });
-      },
-    },
-
-    get_gate_verdict: {
-      annotations: { title: 'Check the gate once', ...READS },
-      outputSchema: {
-        type: 'object',
-        properties: {
-          status: {
-            type: 'string',
-            enum: ['pending', ...GATE_STATUS_VALUES],
-          },
-          deliveryId: { type: ['string', 'null'] },
-          summary: { type: 'string' },
-          access: { type: 'string' },
-          version: { type: 'string' },
-          green: { type: 'boolean' },
-          lane: { type: 'string', enum: ['preview', 'publish'] },
-          ranAt: { type: 'string' },
-          report: { type: 'string' },
-          gateStatus: { type: 'string' },
-          previewPassed: { type: 'boolean' },
-          retryAfterSeconds: {
-            type: 'number',
-            description:
-              'Informational delay before a later creator-led run checks again. stop:true takes priority in this run.',
-          },
-          stop: { type: 'boolean', description: 'When true, stop this agent run immediately.' },
-          reason: { type: 'string' },
-          ...WARNINGS_PROP,
-        },
-        required: ['status', 'deliveryId', 'summary', 'access', 'stop'],
-      },
-      description:
-        'One-shot check of the gate verdict for a delivery (default: latest); this is not a polling or waiting tool. ' +
-        'Preview lane: preview_passed / preview_failed ' +
-        '(does not end the round). Publish lane: green / red / kit_outdated — only green ends the round. ' +
-        'Verdicts typically land in 2–5 minutes. When status=pending and deliveryId is set, the result has stop:true: ' +
-        'STOP this run immediately and let Studio show the eventual result. A pending result with deliveryId:null means ' +
-        'you checked before delivering: stop is false, so continue building and call submit_sources instead of checking again. ' +
-        'retryAfterSeconds is only for a later creator-led run checking a delivered gate. Repeated checks trigger warnings.code=gate_poll_backoff. ' +
-        'kit_outdated is terminal — stop polling, re-run get_kit, then submit_sources({ fromLatestDelivery: true, mode, kitEngineRef }) ' +
-        '(same mode as the refused delivery; omit mode only to reuse that lane; do not re-upload the whole tree; do not wait for green/red). ' +
-        'Terminal receipt: still readable after the round closes ' +
-        "when your capability's generation owns that delivery (generation may be exactly one behind current), " +
-        'so the verdict stays readable if the round closes between polls. ' +
-        'Expiry still applies. Wait for publish green before considering the round done. ' +
-        BEHAVIOURAL_CONTRACT,
-      inputSchema: {
-        type: 'object',
-        properties: {
-          sessionKey: SESSION_KEY_PROP,
-          deliveryId: { type: 'string', description: "Delivery version id; default is the job's latest." },
-        },
-        required: [],
-      },
-      handler: async (args, ctx) => {
-        const auth = await resolveAuth(ctx, args, { allowTerminalReceipt: true });
-        if (!('channelToken' in auth)) return auth;
-
-        const deliveryId =
-          typeof args.deliveryId === 'string' && args.deliveryId.trim() ? args.deliveryId.trim() : null;
-        const path = deliveryId
-          ? `${AGENT_CHANNEL_ROUTES.GATE}?version=${encodeURIComponent(deliveryId)}`
-          : AGENT_CHANNEL_ROUTES.GATE;
-        const res = await injectChannel(ctx.request, 'GET', path, auth.channelToken);
-        const body = res.json() as Record<string, unknown> & { error?: string; status?: string };
-        if (res.statusCode !== 200) {
-          return toolErr(body.error ?? `gate verdict failed (${res.statusCode})`);
-        }
-        if (body.status === 'pending' && typeof body.deliveryId === 'string') {
-          return toolOk({
-            ...body,
-            summary:
-              'gate is still running — STOP this agent run now; do not call get_gate_verdict or any other tool again. Studio will show the eventual result.',
-            stop: true,
-            reason: 'gate_pending',
-          });
-        }
-        if (body.status === 'pending') {
-          return toolOk({
-            ...body,
-            summary:
-              'nothing has been delivered yet — continue building and call submit_sources; do not call get_gate_verdict again before a delivery',
-            stop: false,
-            reason: 'no_delivery',
-          });
-        }
-        if (body.status === 'green') {
-          return toolOk({ ...body, stop: true, reason: 'gate_green' });
-        }
-        return toolOk({ ...body, stop: false });
-      },
-    },
-
-    /**
-     * App-only companion to get_gate_media, for the round view.
-     *
-     * The agent cannot screenshot a game it has no browser to run — observed repeatedly,
-     * and correctly: the alternative is drawing an approximation and calling it a
-     * capture. The gate does run a browser and captures on both lanes, so its frames are
-     * the only honest picture of the build. Reaching them through the agent means waiting
-     * for a verdict, which the workflow tells it not to do. The card is already waiting,
-     * so it fetches them itself.
-     *
-     * Unlike get_gate_media this returns the bytes inside structuredContent rather than
-     * as image blocks: a view needs data URIs, and there is no model context to protect
-     * because the model never sees this call.
-     */
-    get_round_media: {
-      annotations: { title: 'Gate frames for the round view', ...READS },
-      outputSchema: {
-        type: 'object',
-        properties: {
-          available: { type: 'boolean' },
-          deliveryId: { type: ['string', 'null'] },
-          lane: { type: ['string', 'null'] },
-          frames: {
-            type: 'array',
-            items: {
-              type: 'object',
-              properties: {
-                file: { type: 'string' },
-                name: { type: 'string' },
-                png: { type: 'string', description: 'base64 PNG' },
-              },
-            },
-          },
-          video: { type: ['object', 'null'], properties: { file: { type: 'string' }, url: { type: 'string' } } },
-          framesOmitted: { type: 'number', description: 'Frames the gate captured but this reply could not carry.' },
-          reason: { type: 'string', description: 'Why there is nothing to show, when there is nothing to show.' },
-        },
-        required: ['available', 'frames'],
-      },
-      description:
-        "The gate's own frames for a delivery, as base64 PNGs the round view can render, plus the gameplay " +
-        'video link when one exists. Read-only, app-only, and presence-neutral like get_round_status.',
-      inputSchema: {
-        type: 'object',
-        properties: {
-          sessionKey: SESSION_KEY_PROP,
-          deliveryId: { type: 'string', description: "Delivery version id; default is the job's latest." },
-        },
-        required: [],
-      },
-      handler: async (args, ctx) => {
-        const auth = await resolveAuth(ctx, args, { allowTerminalReceipt: true });
-        if (!('channelToken' in auth)) return auth;
-
-        const deliveryId =
-          typeof args.deliveryId === 'string' && args.deliveryId.trim() ? args.deliveryId.trim() : null;
-        const query = new URLSearchParams({ frames: 'all' });
-        if (deliveryId) query.set('version', deliveryId);
-        const res = await injectChannel(
-          ctx.request,
-          'GET',
-          `${AGENT_CHANNEL_ROUTES.MEDIA}?${query.toString()}`,
-          auth.channelToken,
-        );
-        const body = res.json() as Record<string, unknown> & {
-          error?: string;
-          frames?: Array<{ file?: string; name?: string; png?: string }>;
-          video?: unknown;
-          framesOmitted?: number;
-          // The lane lives under the verdict, not at the top level: which run captured
-          // these frames is a property of the gate that took them.
-          gate?: { lane?: unknown };
-          reason?: string;
-        };
-        if (res.statusCode !== 200) {
-          return toolErr(body.error ?? `gate media failed (${res.statusCode})`);
-        }
-
-        // Bounded again on our side, in base64 rather than in bytes.
-        //
-        // The channel already caps what it inlines, but it caps *decoded* size for the
-        // sake of a model's context. What binds here is different: these frames ride a
-        // JSON-RPC postMessage through the host into a sandboxed iframe, base64 and all,
-        // which is a third larger than the channel measured. So the channel's ceiling is
-        // not this one's, and a run that squeaks under it can still be too much to send.
-        // Whatever is dropped is counted, never silently lost.
-        let budget = ROUND_MEDIA_BYTE_BUDGET;
-        const frames: Array<{ file: string; name: string; png: string }> = [];
-        let omitted = typeof body.framesOmitted === 'number' ? body.framesOmitted : 0;
-        for (const frame of body.frames ?? []) {
-          const png = typeof frame.png === 'string' ? frame.png : '';
-          if (!png) continue;
-          if (frames.length >= ROUND_MEDIA_MAX_FRAMES || png.length > budget) {
-            omitted += 1;
-            continue;
-          }
-          budget -= png.length;
-          frames.push({ file: String(frame.file ?? ''), name: String(frame.name ?? frame.file ?? ''), png });
-        }
-
-        const lane = typeof body.gate?.lane === 'string' ? body.gate.lane : null;
-        return toolOk({
-          available: frames.length > 0 || Boolean(body.video),
-          deliveryId: (body.deliveryId as string | undefined) ?? deliveryId,
-          lane,
-          frames,
-          video: (body.video as Record<string, unknown> | undefined) ?? null,
-          ...(omitted > 0 ? { framesOmitted: omitted } : {}),
-          // Why there is nothing to show, when there is nothing to show. The card can
-          // say "the gate stored no media" instead of rendering an empty strip.
-          ...(frames.length === 0 && !body.video && body.reason ? { reason: body.reason } : {}),
-        });
-      },
-    },
-
-    get_gate_media: {
-      annotations: { title: "Fetch the gate's screenshots and video", ...READS },
-      outputSchema: {
-        type: 'object',
-        properties: {
-          available: { type: 'boolean' },
-          deliveryId: { type: ['string', 'null'] },
-          screenshots: {
-            type: 'array',
-            items: {
-              type: 'object',
-              properties: { file: { type: 'string' }, url: { type: 'string' } },
-              required: ['file', 'url'],
-            },
-          },
-          video: {
-            type: ['object', 'null'],
-            properties: { file: { type: 'string' }, url: { type: 'string' } },
-          },
-          openingShot: {
-            type: 'object',
-            properties: { file: { type: 'string' }, attached: { type: 'boolean' } },
-            required: ['file', 'attached'],
-          },
-          access: { type: 'string' },
-        },
-        required: ['available', 'deliveryId'],
-      },
-      description:
-        'Fetch the media the gate itself produced for a delivery (default: latest). Screenshots come back ' +
-        'BOTH as attached images (no fetching needed — use these) and as short-lived signed URLs; the ' +
-        'gameplay MP4 is a URL only. ' +
-        'Use it when you cannot run the game yourself — look at the attached frames for visual defects ' +
-        '(blank canvas, missing sprites) before resubmitting, and show them to the creator. ' +
-        'frames=opening (default) attaches one frame; frames=all attaches up to 3; frames=none skips them ' +
-        'when you only want the URLs. ' +
-        'If your client cannot open URLs, do not try and do not report the video as broken — hand the link ' +
-        'to the creator, who can, and describe the game from the attached frames. ' +
-        'Read-only over the gate run that already happened; it never triggers a build, and media exists only ' +
-        'after a delivery has been gated. Terminal receipt: like get_gate_verdict, the latest delivery stays ' +
-        'readable after green closes the round. ' +
-        BEHAVIOURAL_CONTRACT,
-      inputSchema: {
-        type: 'object',
-        properties: {
-          sessionKey: SESSION_KEY_PROP,
-          deliveryId: { type: 'string', description: "Delivery version id; default is the job's latest." },
-          frames: {
-            type: 'string',
-            enum: ['opening', 'all', 'none'],
-            description:
-              'How many screenshots to attach as images: opening (default, one), all (up to 3), none (URLs only).',
-          },
-        },
-        required: [],
-      },
-      handler: async (args, ctx) => {
-        const auth = await resolveAuth(ctx, args, { allowTerminalReceipt: true });
-        if (!('channelToken' in auth)) return auth;
-
-        const deliveryId =
-          typeof args.deliveryId === 'string' && args.deliveryId.trim() ? args.deliveryId.trim() : null;
-        const frames = args.frames === 'all' || args.frames === 'none' ? args.frames : 'opening';
-        const query = new URLSearchParams({ frames });
-        if (deliveryId) query.set('version', deliveryId);
-        const res = await injectChannel(
-          ctx.request,
-          'GET',
-          `${AGENT_CHANNEL_ROUTES.MEDIA}?${query.toString()}`,
-          auth.channelToken,
-        );
-        const body = res.json() as Record<string, unknown> & {
-          error?: string;
-          frames?: Array<{ file?: string; name?: string; png?: string }>;
-        };
-        if (res.statusCode !== 200) {
-          return toolErr(body.error ?? `gate media failed (${res.statusCode})`);
-        }
-        // Frames travel as MCP image blocks, never inside the JSON body: base64 in the
-        // text content would double the cost and no client renders it. The structured
-        // half keeps the names so the model can talk about what it was shown.
-        const { frames: inlineFrames, ...rest } = body;
-        const attached = (inlineFrames ?? []).filter((frame) => typeof frame.png === 'string' && frame.png);
-        const structured = {
-          ...rest,
-          frames: attached.map((frame) => ({ file: frame.file, name: frame.name, attached: true })),
-        };
-        const result = toolOk(structured);
-        for (const frame of attached) {
-          result.content.push({ type: 'image', data: frame.png as string, mimeType: 'image/png' });
-        }
-        return result;
-      },
-    },
+    ...createRoundCardTools({ resolveAuth, injectChannel, store, now }),
+    ...createGateMediaTools({ resolveAuth, injectChannel }),
 
     ...createInboxTools({ resolveAuth, injectChannel, writePiggyback }),
   };
